+++ lootmarshal/cli/commands/__init__.py
@@ -1,3 +1,3 @@
 from .connect import connect_app
-from .secrets import secret_app
 from .creds import creds_app
+from .secrets import secret_app