+++ lootmarshal/server/handlers/handlercontext.py
@@ -1,4 +1,5 @@
 from abc import ABC, abstractmethod
+
 from fastapi import HTTPException, status
 
 from ...settings import _ENV_FILE
@@ -12,9 +13,8 @@ class HandlerContext:
     @classmethod
     def setContext(cls, ctx: str) -> str:
         from .azurehandler import AzureHandler
-        from .filehandler import FileHandler
 
-        _handlers = {"azure": AzureHandler, "file": FileHandler}
+        _handlers = {"azure": AzureHandler}
         try:
             cls.handler = _handlers[ctx](_ENV_FILE)
             return f"Connection to {ctx} successful!"
@@ -31,22 +31,36 @@ class HandlerContext:
 
 
 class Handler(ABC):
+
+    secret_index = []
+
     @abstractmethod
     async def validate(self):
         pass
 
     @abstractmethod
-    async def get_store(self):
+    async def build_secret_index(self):
+        """Create an dict index of secret tags where key is secret name and values are tuple of tag key/value.
+        Should be called with create_task during handler initialization and assigned to handler secret_index property.
+
+        Example: {'name': ('tag1', 'value1')}
+        """
         pass
 
     @abstractmethod
-    async def read_secret(self, name):
+    async def write_file(self, directory: str, name: str, file: bytes):
+        """Writes a file to the handler's fileclient.
+        Return Exception if unsuccessful."""
+        pass
+
+    @abstractmethod
+    async def read_secret(self, name: str):
         """Returns a secret from the handler's secretclient.
         Return Exception if unsuccessful."""
         pass
 
     @abstractmethod
-    async def write_secret(self, name, value, content_type):
+    async def write_secret(self, name: str, value: str, content_type: str, tags: dict):
         """Writes a secret to the handler's secretclient.
         Return Exception if unsuccessful."""
         pass
@@ -56,3 +70,6 @@ class Handler(ABC):
         """Returns all secrets from the handler's secretclient.
         Return Exception if unsuccessful."""
         pass
+
+    async def search_secret_index(self, search: str):
+        return [s for s in self.secret_index if search in str(s)]