+++ lootmarshal/server/models/models.py
@@ -1,8 +1,10 @@
+from typing import Optional
+
 from pydantic import BaseModel
 
 
 class SecretModel(BaseModel):
     name: str
     value: str
-    content_type: str
-    tags: dict = {}
+    content_type: Optional[str] = ...
+    tags: Optional[dict]