+++ lootmarshal/server/routers/connect.py
@@ -1,4 +1,5 @@
 from fastapi import APIRouter, HTTPException, status
+
 from ..handlers.handlercontext import HandlerContext as HC
 
 router = APIRouter()