+++ lootmarshal/settings/tags.py
@@ -1,6 +1,0 @@
-TAGS = [
-    "domain",
-    "hostname",
-    "tenant",
-    "username",
-]